@@ -1,6 +1,19 @@
 [package]
 name = "regex_engine"
 version = "0.1.0"
-edition = "2021"
+edition = "2024"
 
-[dependencies]
+[dev-dependencies]
+criterion = { version = "0.5", features = ["html_reports"] }
+
+[[bench]]
+name = "thompson_benchmark"
+harness = false
+
+[[bench]]
+name = "glushkov_benchmark"
+harness = false
+
+[[bench]]
+name = "rust_regex_benchmark"
+harness = false
